+++ source/globalCommands.py
@@ -2,7 +2,7 @@
 # A part of NonVisual Desktop Access (NVDA)
 # This file is covered by the GNU General Public License.
 # See the file COPYING for more details.
-# Copyright (C) 2006-2024 NV Access Limited, Peter Vágner, Aleksey Sadovoy, Rui Batista, Joseph Lee,
+# Copyright (C) 2006-2025 NV Access Limited, Peter Vágner, Aleksey Sadovoy, Rui Batista, Joseph Lee,
 # Leonard de Ruijter, Derek Riemer, Babbage B.V., Davy Kager, Ethan Holliger, Łukasz Golonka, Accessolutions,
 # Julien Cochuyt, Jakub Lukowicz, Bill Dengler, Cyrille Bougot, Rob Meredith, Luke Davis,
 # Burman's Computer and Education Ltd.
@@ -3221,25 +3221,6 @@ class GlobalCommands(ScriptableObject):
 			# Translators: presented when toggled.
 			state = _("focus moves navigator object on")
 			config.conf["reviewCursor"]["followFocus"] = True
-		ui.message(state)
-
-	@script(
-		description=_(
-			# Translators: Input help mode message for toggle auto focus focusable elements command.
-			"Toggles on and off automatic movement of the system focus due to browse mode commands",
-		),
-		category=inputCore.SCRCAT_BROWSEMODE,
-		gesture="kb:NVDA+8",
-	)
-	def script_toggleAutoFocusFocusableElements(self, gesture):
-		if config.conf["virtualBuffers"]["autoFocusFocusableElements"]:
-			# Translators: presented when toggled.
-			state = _("Automatically set system focus to focusable elements off")
-			config.conf["virtualBuffers"]["autoFocusFocusableElements"] = False
-		else:
-			# Translators: presented when toggled.
-			state = _("Automatically set system focus to focusable elements on")
-			config.conf["virtualBuffers"]["autoFocusFocusableElements"] = True
 		ui.message(state)
 
 	# added by Rui Batista<ruiandrebatista@gmail.com> to implement a battery status script